+++ R/tools.R
@@ -27,22 +27,25 @@ contain <- function(d, pattern = "NDVI|EVI") {
     names(d) %>% .[grep(pattern, .)]
 }
 
-listk <- function(...){
+listk <- function(...) {
     # get variable names from input expressions
-    cols <- as.list(substitute(list(...)))[-1]
-    vars <- names(cols)
-    Id_noname <- if (is.null(vars)) seq_along(cols) else which(vars == "")
+    vars <- as.list(substitute(list(...)))[-1] # the first element is `list`
+    names <- names(vars)
 
+    Id_noname <- if (is.null(names)) seq_along(vars) else which(names == "")
     if (length(Id_noname) > 0)
-      vars[Id_noname] <- sapply(cols[Id_noname], deparse)
-    # ifelse(is.null(vars), Id_noname <- seq_along(cols), Id_noname <- which(vars == ""))
-    x <- setNames(list(...), vars)
-    return(x)
+        names[Id_noname] <- sapply(vars[Id_noname], deparse)
+    setNames(list(...), names)
+}
+
+export2glob <- function(...) {
+    list2env(listk(...), envir = .GlobalEnv)
+    invisible()
 }
 
 #' melt_list
 #' @keywords internal
-#' @export 
+#' @export
 melt_list <- function (list, var.name = "variable", na.rm = TRUE, ...)
 {
     if (is.null(names(list)))