+++ NEWS.md
@@ -1,8 +1,11 @@
 # phenofit 0.3.4 (2021-11-14)
 
 - Fix the bug of `findpeaks`, which lead to sharp changed growing season failed to detect.
-- Fix the bug of `PhenoKlos` where `minpeakheight` not work in previous version.
-
+- Fix the bug of `PhenoKlos`, where `minpeakheight` not work in previous version.
+- Fix the bug of `check_season_dt`, where `peak` might be able to greater than `end`.
+- Remove the parameter `check_season_dt` in `removeClosedExtreme`, which might eliminate
+  good extreme values.
+  
 # phenofit 0.3.3 (2021-10-26)
 
 - Fix typo error in curvefits' document.